First attempt at POC for "signal handlers can subvert the DDC".
This is very inelegant as it has to move the stack in order to set a DDC
that still allows the program to run, and it doesn't even put the stack
back, instead `exit(0)`ing. It does, however, show that a DDC-restricted
compartment can install a signal handler that will later be called with
whatever DDC happens to be present at the time the signal is handled --
in other words, that a DDC-restricted compartment can use signal
handlers to break out of itself.