Fix stack manipulation.
As it was, the C compiler would attempt to load the value of `old_stack`
to pass into the assembly block, but this load would be sp-based and
because we'd messed with sp, it would load uninitialised data, which the
assembly then moved into sp. This manifested as a bus error, since (in
my tests) that uninitialised data was unaligned.

The fix is to put the whole sequence inside a contiguous assembly block,
which the C compiler won't interfere with. Expressing that inline is
noisy (because it'd have to explicitly list all caller-saved registers
as clobbers) so I moved it to a separate veneer function.

My hope is that the result is still clear, and doesn't cause a
distraction from the point of the example.

This also builds with -O2 because it makes the (dis)assembly easier to
follow.